Avoid PostgreSQL & Redis restarts
Presumably, these restarts were added as a workaround for the fact that
the Rocket network interface may not become available immediately after
a reboot.  In that case Postgres and Redis would start but not listen on
the desired interface.  The first time a Rocket container is started,
the interface will be created; restarting the services afterwards would
ensure that they start listening on the interface.

Obviously, automatically restarting these services every time an
instance is created is not ideal.

The alternative is to configure a network interface according to the
README.md (tap0) which ensures the Interface is available early on.

This patch simply removes the restart tasks.  The README already
contains the necessary information.

See also
https://github.com/OpenSlides/openslides-multiinstance-backend/issues/23.